% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H3K4me3_total.r
\docType{data}
\name{H3K4me3_total}
\alias{H3K4me3_total}
\title{Total human H3K4me3 peaks.}
\format{A data frame with 182470 rows and 3 variables:}
\usage{
data(H3K4me3_total)
}
\description{
A dataset containing all proximal and distal peak coordinates (chromosome number, peak start position, peak end position) for H3K4me3 from ENCODE and ROADMAP Epigenomics.
}
\keyword{datasets}
