% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H3K9ac_total.r
\docType{data}
\name{H3K9ac_total}
\alias{H3K9ac_total}
\title{Total human H3K9ac peaks.}
\format{A data frame with 91233 rows and 3 variables:}
\usage{
data(H3K9ac_total)
}
\description{
A dataset containing all proximal and distal peak coordinates (chromosome number, peak start position, peak end position) for H3K9ac from ENCODE and ROADMAP Epigenomics.
}
\keyword{datasets}
