% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txn_total.r
\docType{data}
\name{txn_total}
\alias{txn_total}
\title{Total human transcription factor peaks.}
\format{A data frame with 210065 rows and 3 variables:}
\usage{
data(txn_total)
}
\description{
A dataset containing all peak coordinates (chromosome number, peak start position, peak end position) for transcription factors from ENCODE.
}
\keyword{datasets}
